Refactor: RoutesDelegatorFactory.php to RouteDelegatorFactory.php
